Air Quality Index Prediction
The objective of this study is to utilize machine learning methods to predict the Air Quality Index (AQI) based on the level of pollutants such as PM2.5, PM10, N2O, CO, O3, and their interaction with temperature and humidity. Performance of the four models, that is, Linear Regression, Gradient Boosting and Random Forest Regression, in predicting the AQI values are compared. The model that was able to achieve the highest R^2 scores was able to reliably handle the non-linear characteristics of the pollutants and the AQI making it the most accurate model for predictions.